feat: add Catalog gRPC Service

+++ src/Catalog.API/Application/Domain/Catalogs/Product.cs
@@ -12,7 +12,7 @@ public class Product : AuditableEntity, IHasDomainEvent
 
     public string Description { get; set; } = string.Empty;
 
-    public decimal Price { get; set; } = 0;
+    public double Price { get; set; } = 0;
 
     public ProductType Type { get; set; }
