+++ src/Catalog.API/Application/Features/Products/Shared/Dtos/ProductDetailResponse.cs
@@ -2,4 +2,4 @@ using CatalogAPI.Application.Domain.Catalogs;
 
 namespace CatalogAPI.Application.Features.Products.Shared.Dtos;
 
-public record ProductDetailResponse(string Id, string Name, string Description, decimal Price, ProductType Type, ProductStatus Status, List<ImageDetailResponse>? Images);
+public record ProductDetailResponse(string Id, string Name, string Description, double Price, ProductType Type, ProductStatus Status, List<ImageDetailResponse>? Images);
